UPDATE: Moved billing to shared
