+++ frontend/src/routes/(protected)/billing/$billingId.tsx
@@ -1,4 +1,4 @@
-import BillingIdPage from '@/pages/admin/billing/$billingId'
+import BillingIdPage from '@/pages/shared/billing/$billingId'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/(protected)/billing/$billingId')({
